Committed version of my java project (star name is "Stay healthy") >> commit: Converted project into a maven supported one. Added a few dependencies
