+++ src/main/java/stay_healthy/GeneralView.fxml
