+++ src/main/java/stay_healthy/Iterator.java
