+++ src/main/java/stay_healthy/LoginController.java
@@ -52,7 +52,6 @@ public class LoginController implements Initializable{
     PreparedStatement preparedStatement = null;
     ResultSet resultSet = null;
 
-    @Override
     public void initialize(URL location, ResourceBundle resources)
     {
         if(con==null)
