+++ src/main/java/stay_healthy/SeeDetailed.fxml
